Check for presence of chunk and section before accessing block data.
If a block is requested from a nonexistent section or chunk, the code
panics.  The code now checks for the presence of a section or chunk
and reports an error.  Eventually the setters will create the
necessary section or chunk which will be helpful when building worlds
the hard way as well as handling tests.