(Mostly) Functional bathy module with tests.
This module uses the ComputeProximity algorithm from the GDAL library
to generate the bathy data.  The Go bindings fail to pass the progress
function properly between the Go code and the C code, causing errors
to be thrown.  The workaround is to build the GDAL library from source
after removing the references to the progress functions in
alg/gdalproximity.cpp and then copy those libraries into place.